requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
meshio==5.3.5
numpy==2.3.1
pillow==11.3.0
pydelatin==0.3.0
Pygments==2.19.2
pypng==0.20220715.0
rich==14.0.0
trimesh==4.7.1
